+++ CHANGELOG.md
@@ -13,3 +13,5 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Changed
 - Run math function asynchronous
 - Disable directly solving equations without command sent to groups
+- Fix not shortening solutions form the calculator
+- Fix message too long for Telegram, for too long solutions from the calculator